authorization_service/requirements.txt
Flask==2.0.1
PyJWT==2.8.0
Werkzeug==2.3.3
requests==2.26.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2==2.9.6
